+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+-   `Tooltip`: fix re-render errors when removing the label.
+
 ## [3.9.0][] - 2024-09-03
 
 ### Added